% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-table.R
\name{cross_table}
\alias{cross_table}
\title{Two Way Tables}
\usage{
cross_table(var1, var2)
}
\arguments{
\item{var1}{First categorical variable}

\item{var2}{Second categorical variable}
}
\value{
\code{cross_table} returns an object of class \code{"cross_table"}.
An object of class \code{"cross_table"} is a list containing at least the
following components:

obs number of observations

var2_levels levels of the second categorical variable
}
\description{
\code{cross_table} creates two way tables of categorical
variables. The tables created can be visualized as barplots and mosaicplots.
}
\details{
cross_table is a generic function
}
\examples{
k <- cross_table(mtcars$cyl, mtcars$gear)

# bar plots
barplot(k)
barplot(k, beside = TRUE)
barplot(k, proportional = TRUE)

# mosaic plots
mosaicplot(k)
}

