% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testlib.R
\docType{package}
\name{testlib}
\alias{testlib}
\alias{testlib-package}
\title{testlib: A package for cross tabulation of data}
\description{
The testlib package provides functions for cross tabulation of categorical data
and visulize the cross tables using bar and mosaic plots
}
\section{testlib functions}{

cross_table
}

